# docs/swagger.yaml
definitions:
  app.SoftwarePkgApplicationDTO:
    properties:
      desc:
        type: string
      license:
        type: string
      platform:
        type: string
      reason:
        type: string
      sig:
        type: string
      source_code:
        type: string
    type: object
  app.SoftwarePkgBasicInfoDTO:
    properties:
      applied_at:
        type: string
      id:
        type: string
      importer:
        type: string
      phase:
        type: string
      pkg_name:
        type: string
      repo_link:
        type: string
    type: object
  app.SoftwarePkgReviewCommentDTO:
    properties:
      author:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
    type: object
  app.SoftwarePkgReviewDTO:
    properties:
      application:
        $ref: '#/definitions/app.SoftwarePkgApplicationDTO'
      applied_at:
        type: string
      approved_by:
        items:
          type: string
        type: array
      comments:
        items:
          $ref: '#/definitions/app.SoftwarePkgReviewCommentDTO'
        type: array
      id:
        type: string
      importer:
        type: string
      phase:
        type: string
      pkg_name:
        type: string
      rejected_by:
        items:
          type: string
        type: array
      repo_link:
        type: string
    type: object
  app.SoftwarePkgsDTO:
    properties:
      pkgs:
        items:
          $ref: '#/definitions/app.SoftwarePkgBasicInfoDTO'
        type: array
      total:
        type: integer
    type: object
  app.TranslatedReveiwCommentDTO:
    properties:
      content:
        type: string
    type: object
  controller.ResponseData:
    properties:
      code:
        type: string
      data: {}
      msg:
        type: string
    type: object
  controller.reviewCommentRequest:
    properties:
      comment:
        type: string
    required:
    - comment
    type: object
  controller.softwarePkgRequest:
    properties:
      desc:
        type: string
      license:
        type: string
      pkg_name:
        type: string
      platform:
        type: string
      reason:
        type: string
      sig:
        type: string
      source_code:
        type: string
    required:
    - desc
    - license
    - pkg_name
    - platform
    - reason
    - sig
    - source_code
    type: object
info:
  contact: {}
paths:
  /v1/softwarepkg:
    get:
      consumes:
      - application/json
      description: list software packages
      parameters:
      - description: importer of the softwarePkg
        in: query
        name: importer
        type: string
      - description: phase of the softwarePkg
        in: query
        name: phase
        type: string
      - description: count per page
        in: query
        name: count_per_page
        type: integer
      - description: page num which starts from 1
        in: query
        name: page_num
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.SoftwarePkgsDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: list software packages
      tags:
      - SoftwarePkg
    post:
      consumes:
      - application/json
      description: apply a new software package
      parameters:
      - description: body of applying a new software package
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/controller.softwarePkgRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: apply a new software package
      tags:
      - SoftwarePkg
  /v1/softwarepkg/{id}:
    get:
      consumes:
      - application/json
      description: get software package
      parameters:
      - description: id of software package
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.SoftwarePkgReviewDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: get software package
      tags:
      - SoftwarePkg
  /v1/softwarepkg/{id}/review/abandon:
    put:
      consumes:
      - application/json
      description: abandon software package
      parameters:
      - description: id of software package
        in: path
        name: id
        required: true
        type: string
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: abandon software package
      tags:
      - SoftwarePkg
  /v1/softwarepkg/{id}/review/approve:
    put:
      consumes:
      - application/json
      description: approve software package
      parameters:
      - description: id of software package
        in: path
        name: id
        required: true
        type: string
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: approve software package
      tags:
      - SoftwarePkg
  /v1/softwarepkg/{id}/review/comment:
    post:
      consumes:
      - application/json
      description: create a new software package review comment
      parameters:
      - description: body of creating a new software package review comment
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/controller.reviewCommentRequest'
      - description: id of software package
        in: path
        name: id
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: create a new software package review comment
      tags:
      - SoftwarePkg
  /v1/softwarepkg/{id}/review/reject:
    put:
      consumes:
      - application/json
      description: reject software package
      parameters:
      - description: id of software package
        in: path
        name: id
        required: true
        type: string
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: reject software package
      tags:
      - SoftwarePkg
  /v1/softwarepkg/translate/review/comment:
    get:
      consumes:
      - application/json
      description: get translate review comment
      parameters:
      - description: pkg_id of software package
        in: query
        name: pkg_id
        required: true
        type: string
      - description: comment_id of review comment
        in: query
        name: comment_id
        required: true
        type: string
      - description: language
        in: query
        name: language
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.TranslatedReveiwCommentDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: get translate review comment
      tags:
      - SoftwarePkg
swagger: "2.0"